refactor: 각 message에 userId를 추가하도록 수정, rq를 사용하도록 변경

+++ src/main/java/com/example/MnM/boundedContext/chat/controller/RoomController.java
@@ -1,7 +1,6 @@
 package com.example.MnM.boundedContext.chat.controller;
 
 import com.example.MnM.base.rq.Rq;
-import com.example.MnM.boundedContext.chat.dto.ChatRoomDto;
 import com.example.MnM.boundedContext.chat.dto.DeleteRoomDto;
 import com.example.MnM.boundedContext.chat.entity.ChatRoom;
 import com.example.MnM.boundedContext.chat.service.RoomService;
@@ -15,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@PreAuthorize("isAuthenticated()")
 @RequiredArgsConstructor
 @RequestMapping("/chat")
 @Controller
@@ -32,10 +32,9 @@ public class RoomController {
     }
 
     @PostMapping("/create/room")
-    public String createRoom(Model model) {
-//        Long roomId = roomService.createRoom(rq.getMember().getUsername());
-        String roomId = roomService.createRoom();
-        model.addAttribute("roomId",roomId);
+    public String createRoom() {
+        String roomId = roomService.createRoom(rq.getMember().getId(),rq.getMember().getUsername());
+
         return rq.redirectWithMsg("/chat/room/%s".formatted(roomId),"채팅 방이 생성되었습니다.");
     }
 
@@ -51,19 +50,19 @@ public class RoomController {
     @DeleteMapping("/room/delete")
     public ResponseEntity<String> deleteRoom(DeleteRoomDto deleteRoomDto) {
 
-//        Long memberId = rq.getMember().getId();
-//
-//        if (!isRoomOwner(deleteRoomDto, memberId)) {
-//            return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
-//        }
+        Long memberId = rq.getMember().getId();
+
+        if (!isRoomOwner(deleteRoomDto, memberId)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
 
         roomService.deleteRoom(deleteRoomDto.getRoomId());
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-//    private boolean isRoomOwner(DeleteRoomDto deleteRoomDto, Long memberId) {
-//        return deleteRoomDto.getUserId().equals(memberId);
-//    }
+    private boolean isRoomOwner(DeleteRoomDto deleteRoomDto, Long memberId) {
+        return deleteRoomDto.getUserId().equals(memberId);
+    }
 
 
 }
